# app/Http/Controllers/Api/V1/AuthController.php
<?php

namespace App\Http\Controllers\Api\V1;

use App\Http\Controllers\Concerns\ResolvesDriver;
use App\Http\Controllers\Controller;
use App\Http\Requests\Api\V1\OtpRequestRequest;
use App\Http\Requests\Api\V1\OtpVerifyRequest;
use App\Http\Requests\Api\V1\UpdateDriverPhotoRequest;
use App\Http\Requests\Api\V1\UpdatePushTokenRequest;
use App\Http\Resources\DriverResource;
use App\Models\Driver;
use App\Services\Auth\OtpService;
use App\Settings\OtpSettings;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Storage;
use Illuminate\Validation\ValidationException;
use Symfony\Component\HttpFoundation\StreamedResponse;

class AuthController extends Controller
{
    use ResolvesDriver;

    public function __construct(private OtpService $otpService) {}

    /**
     * Demander un code OTP
     *
     * Envoie un code à 6 chiffres (valable 5 minutes) par SMS ou WhatsApp. Le
     * conducteur doit être pré-enregistré par le back-office : l'application
     * mobile ne crée pas de compte.
     *
     * Limité à 3 envois par tranche de 10 minutes et par numéro.
     *
     * En développement et en test (`WIGO_OTP_EXPOSE_CODE`), la réponse contient
     * en plus le champ `code` : il est toujours absent en production.
     *
     * @response array{
     *     message: string,
     *     data: array{
     *         channel: 'sms'|'whatsapp',
     *         expires_at: string,
     *         code?: string,
     *     },
     * }
     */
    public function requestOtp(OtpRequestRequest $request): JsonResponse
    {
        $driver = $this->findByPhone($request->string('phone')->toString());

        $otpCode = $this->otpService->send($driver, $request->channel(), $request->ip());

        $payload = [
            'channel' => $otpCode->channel->value,
            'expires_at' => $otpCode->expires_at->toIso8601String(),
        ];

        $plainCode = $this->otpService->lastPlainCode();

        if ($plainCode !== null) {
            $payload['code'] = $plainCode;
        }

        return $this->okApiResponse($payload, __('otp.sent', ['minutes' => app(OtpSettings::class)->ttl_minutes]));
    }

    /**
     * Vérifier l'OTP et obtenir un jeton
     *
     * Délivre un jeton Sanctum portant l'habilitation `mobile:*`, à placer dans
     * l'en-tête `Authorization: Bearer <token>`. Plusieurs codes peuvent être
     * valides simultanément ; une vérification réussie les consomme tous.
     *
     * Après 5 saisies erronées, le numéro est verrouillé 15 minutes.
     *
     * @response array{
     *     message: string,
     *     data: array{
     *         token: string,
     *         driver: DriverResource,
     *         terms: array{current_version: string, accepted: bool},
     *     },
     * }
     */
    public function verifyOtp(OtpVerifyRequest $request): JsonResponse
    {
        $driver = $this->findByPhone($request->string('phone')->toString());

        $this->otpService->verify($driver, $request->string('code')->toString());

        if ($request->filled('terms_version')) {
            $driver->forceFill([
                'terms_version' => $request->string('terms_version')->toString(),
                'terms_accepted_at' => now(),
            ])->save();
        }

        $token = $driver->createToken($request->string('device_name')->toString(), ['mobile:*']);

        return $this->okApiResponse([
            'token' => $token->plainTextToken,
            // `resolve()` : le conducteur reste un objet imbriqué simple, sans
            // seconde enveloppe `{message, data}`.
            'driver' => (new DriverResource($driver->load('vehicle')))->resolve(),
            'terms' => [
                'current_version' => config('wigo.terms_version'),
                'accepted' => $driver->hasAcceptedCurrentTerms(),
            ],
        ]);
    }

    /**
     * Se déconnecter
     *
     * Révoque uniquement le jeton présenté ; les autres appareils du conducteur
     * restent connectés.
     *
     * @response array{message: string, data: array<string, never>}
     */
    public function logout(Request $request): JsonResponse
    {
        $this->driver($request)->currentAccessToken()->delete();

        return $this->okApiResponse([], __('auth.logged_out'));
    }

    /**
     * Profil du conducteur
     *
     * Accessible même lorsque le compte est suspendu, afin que l'application
     * puisse afficher le motif et permettre la déconnexion.
     */
    public function me(Request $request): JsonResponse
    {
        return $this->okApiResponse(new DriverResource($this->driver($request)->load('vehicle')));
    }

    /**
     * Enregistrer le jeton FCM de l'appareil
     *
     * Requis pour recevoir les notifications push (messages data-only).
     *
     * @response array{message: string, data: array<string, never>}
     */
    public function updatePushToken(UpdatePushTokenRequest $request): JsonResponse
    {
        $this->driver($request)->forceFill([
            'fcm_token' => $request->string('fcm_token')->toString(),
        ])->save();

        return $this->okApiResponse([], __('auth.push_token_saved'));
    }

    /**
     * Changer ma photo de profil
     *
     * Remplace la photo du conducteur : le profil est mis à jour aussitôt, il
     * n'y a pas de validation par un agent. L'ancienne photo est supprimée du
     * disque dans la foulée.
     *
     * Accessible même lorsque le compte est suspendu — la photo relève du
     * profil, pas des opérations métier.
     *
     * @response array{
     *     message: string,
     *     data: array{
     *         photo_url: string,
     *     },
     * }
     */
    public function updatePhoto(UpdateDriverPhotoRequest $request): JsonResponse
    {
        $driver = $this->driver($request);
        $previousPath = $driver->photo_url;

        // Disque privé : un portrait est une donnée personnelle, il n'a rien à
        // faire derrière une URL publique devinable (comme les justificatifs
        // CNPS). La lecture passe par `photo`, route signée et temporaire.
        $path = $request->file('photo')->store("driver-photos/{$driver->id}", 'local');

        $driver->forceFill(['photo_url' => $path])->save();

        if ($previousPath !== null && $previousPath !== $path) {
            Storage::disk('local')->delete($previousPath);
        }

        return $this->okApiResponse([
            'photo_url' => DriverResource::photoUrl($driver),
        ], __('api.driver.photo_updated'));
    }

    /**
     * Télécharger ma photo de profil
     *
     * Route signée et temporaire, dont l'URL est fournie par `photo_url` du
     * profil. La signature ne vaut pas autorisation : le conducteur
     * authentifié doit être celui de la photo.
     */
    public function photo(Request $request, Driver $driver): StreamedResponse
    {
        abort_if($driver->id !== $this->driver($request)->id, 403, __('api.forbidden'));
        abort_if($driver->photo_url === null, 404, __('api.driver.photo_missing'));

        $disk = Storage::disk('local');

        abort_unless($disk->exists($driver->photo_url), 404, __('api.driver.photo_missing'));

        return $disk->response($driver->photo_url);
    }

    /**
     * @throws ValidationException si le numéro n'est pas enregistré
     */
    private function findByPhone(string $phone): Driver
    {
        $driver = Driver::where('phone', $phone)->first();

        if ($driver === null) {
            throw ValidationException::withMessages([
                'phone' => [__('otp.unknown_phone')],
            ]);
        }

        return $driver;
    }
}